Fix qci gate emit on windows

+++ src/core4inet/linklayer/filtering/IEEE8021Qci/gate/IEEE8021QciGate.cc
@@ -56,7 +56,7 @@ void IEEE8021QciGate::handleMessage(cMessage *msg)
         else if (this->state == this->State::CLOSED)
         {
             this->bubble("Drop frame");
-            this->emit(frameDroppedSignal, ctrl->getEncapsulatedPacket()->getByteLength());
+            this->emit(frameDroppedSignal, static_cast<unsigned long>(ctrl->getEncapsulatedPacket()->getByteLength()));
             delete ctrl;
             this->numFramesDropped++;
         }
